+++ herdr-agent-quota/tests/install_upgrade.rs
@@ -1,4 +1,6 @@
 #![cfg(unix)]
+
+#![cfg(unix)]
 use std::{fs, os::unix::fs::PermissionsExt, process::Command};
 
 #[test]