+++ herdr-agent-quota/tests/watch_handoff.rs
@@ -1,5 +1,7 @@
 #![cfg(unix)]
 
+#![cfg(unix)]
+
 use std::fs;
 use std::os::unix::fs::PermissionsExt;
 use std::process::{Command, Stdio};